+++ src/Candoumbe.Pipelines/Components/IReportCoverage.cs
@@ -17,7 +17,7 @@ using static Nuke.Common.Tools.ReportGenerator.ReportGeneratorTasks;
 /// <remarks>
 /// This component requires <see href="https://nuget.org/packages/reportgenerator">ReportGenerator tool</see>
 /// </remarks>
-public interface IReportCoverage : INukeBuild, IRequireNuGetPackage
+public interface IReportCoverage : IRequireNuGetPackage, IHaveCoverage
 {
     /// <summary>
     /// The API key used to push code coverage to CodeCov