+++ src/Candoumbe.Pipelines/Components/IReportUnitTestCoverage.cs
@@ -19,22 +19,32 @@ using static Nuke.Common.Tools.ReportGenerator.ReportGeneratorTasks;
 /// </remarks>
 public interface IReportUnitTestCoverage : IReportCoverage, IUnitTest
 {
+    /// <summary>
+    /// Directory where coverage report history files will be pushed
+    /// </summary>
+    public AbsolutePath UnitTestCoverageReportDirectory => CoverageReportDirectory / "unit-tests";
+
+    /// <summary>
+    /// Directory where coverage report history files will be pushed
+    /// </summary>
+    public AbsolutePath UnitTestCoverageReportHistoryDirectory => CoverageReportHistoryDirectory / "unit-tests";
+
     internal sealed Configure<ReportGeneratorSettings> ReportGeneratorSettingsBase => _ => _
         .SetFramework("net5.0")
         .SetReports(UnitTestResultsDirectory / "*.xml")
         .SetReportTypes(ReportTypes.Badges, ReportTypes.HtmlChart, ReportTypes.HtmlInline)
         .WhenNotNull(this.As<IHaveGitRepository>()?.GitRepository,
             (_, repository) => !string.IsNullOrWhiteSpace(repository.Branch)
-                ? _.SetTargetDirectory(CoverageReportDirectory / repository.Branch)
-                    .SetHistoryDirectory(CoverageReportHistoryDirectory / repository.Branch)
+                ? _.SetTargetDirectory(UnitTestCoverageReportDirectory / repository.Branch)
+                    .SetHistoryDirectory(UnitTestCoverageReportHistoryDirectory / repository.Branch)
                     .SetTag(repository.Commit)
-                : _.SetTargetDirectory(CoverageReportDirectory)
-                    .SetHistoryDirectory(CoverageReportHistoryDirectory)
+                : _.SetTargetDirectory(UnitTestCoverageReportDirectory)
+                    .SetHistoryDirectory(UnitTestCoverageReportHistoryDirectory)
                     .SetTag(repository.Commit)
         )
         .When(_ => this.As<IHaveGitRepository>() is null,
-            _ => _.SetTargetDirectory(CoverageReportDirectory)
-                .SetHistoryDirectory(CoverageReportHistoryDirectory));
+            _ => _.SetTargetDirectory(UnitTestCoverageReportDirectory)
+                .SetHistoryDirectory(UnitTestCoverageReportHistoryDirectory));
 
     internal sealed Configure<CodecovSettings> CodeCovSettingsBase => _ => _
         .SetFiles(UnitTestResultsDirectory.GlobFiles("*.xml").Select(x => x.ToString()))
@@ -53,8 +63,8 @@ public interface IReportUnitTestCoverage : IReportCoverage, IUnitTest
        .DependsOn(UnitTests)
        .Requires(() => !ReportToCodeCov || CodecovToken != null)
        .Consumes(UnitTests, UnitTestResultsDirectory / "*.xml")
-       .Produces(CoverageReportDirectory / "*.xml")
-       .Produces(CoverageReportHistoryDirectory / "*.xml")
+       .Produces(UnitTestCoverageReportDirectory / "*.xml")
+       .Produces(UnitTestCoverageReportHistoryDirectory / "*.xml")
        .Executes(() =>
        {
            ReportGenerator(s => s.Apply(ReportGeneratorSettingsBase)